Moved starting logic from main to parser
